// src/lib/showfile/generated/CuelistDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CueDto } from "./CueDto";
import type { CuelistKind } from "./CuelistKind";

export type CuelistDto = { id: number, name: string, kind: CuelistKind, tracking: boolean, cues: Array<CueDto>, };