// src/lib/showfile/generated/CuelistKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChaseDirection } from "./ChaseDirection";

export type CuelistKind = { "kind": "standard" } | { "kind": "chase", bpm: number, direction: ChaseDirection, } | { "kind": "override" } | { "kind": "submaster" } | { "kind": "timecode" };