// src/lib/showfile/generated/PlaybackStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FaderMode } from "./FaderMode";

export type PlaybackStatus = { id: number, cuelistId: number, currentCueNumber: number | null, fader: number, faderMode: FaderMode, paused: boolean, };